requirements.txt
streamlit
azure-ai-inference